uodguapi/requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
autopep8==1.4.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.5.0
discord.py==0.16.0
Django==2.1.7
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-registration==3.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.9.2
future==0.17.1
idna==2.8
jedi==0.13.2
mccabe==0.6.1
multidict==4.5.2
mysql-connector==2.1.6
oauthlib==3.0.1
parso==0.3.3
Pillow==5.4.1
pkg-resources==0.0.0
pluggy==0.8.1
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==3.0.0
pyflakes==2.1.0
pymssql==2.1.4
PyNaCl==1.3.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
snowballstemmer==1.2.1
urllib3==1.24.1
websockets==3.4
